complete error handling store_fund

+++ model/StoreFund.js
@@ -139,12 +139,20 @@ const storeFund = {
     read: (storeIdx) => {
         return new Promise(async (resolve, reject) => {
             const selectStoreFundInfoQuery = `SELECT * FROM ${table} WHERE store_idx = ?`;
-            const selectStoreFundInfoResult = await pool.queryParam_Arr(selectStoreFundInfoQuery, [storeIdx]);
+            const selectStoreFundInfoResult = await pool.queryParam_Arr(selectStoreFundInfoQuery, [storeIdx])
 
             if (!selectStoreFundInfoResult) {
                 resolve({
                     code: statusCode.INTERNAL_SERVER_ERROR,
                     json: authUtil.successFalse(statusCode.INTERNAL_SERVER_ERROR, responseMessage.INTERNAL_SERVER_ERROR)
+                });
+                return;
+            }
+
+            if (selectStoreFundInfoResult[0] == undefined) {
+                resolve({
+                    code : statusCode.BAD_REQUEST,
+                    json : authUtil.successFalse(statusCode.BAD_REQUEST, `존재하지 않는 가게 펀드 정보`)
                 });
                 return;
             }
@@ -157,9 +165,19 @@ const storeFund = {
     },
     update: (storeIdx, customerCount, marginPercent, goalMoney, remaining_days, fund_status ) => {
         return new Promise(async (resolve, reject) => {
+            const storeIdxQuery = `SELECT * FROM ${table} WHERE store_idx = ?`;
+            const storeIdxResult = await pool.queryParam_Arr(storeIdxQuery, [storeIdx]);
+
+            if (storeIdxResult[0] == undefined) {
+                resolve({
+                    code : statusCode.BAD_REQUEST,
+                    json : authUtil.successFalse(statusCode.BAD_REQUEST, `존재하지 않는 가게 펀드 정보(잘못된 인덱스)`)
+                });
+                return;
+            }
             
-            const updateStoreFundInfoQuery = `UPDATE ${table} SET customer_count = ?, margin_percent = ?, goal_money = ?, remaining_days = ?, fund_status = ? WHERE store_idx = ?`;
-            const updateStoreFundInfoResult = await pool.queryParam_Arr(updateStoreFundInfoQuery, [customerCount, marginPercent, goalMoney, storeIdx, remaining_days, fund_status]);
+            const updateStoreFundInfoQuery = `UPDATE ${table} SET customer_count = ?, margin_percent = ?, goal_money = ?, fund_status = ? WHERE store_idx = ?`;
+            const updateStoreFundInfoResult = await pool.queryParam_Arr(updateStoreFundInfoQuery, [customerCount, marginPercent, goalMoney, fund_status, storeIdx]);
 
             if(!updateStoreFundInfoResult){
                 resolve({
@@ -178,6 +196,17 @@ const storeFund = {
 
     delete: (storeIdx) => {
         return new Promise(async (resolve, reject) => {
+            const storeIdxQuery = `SELECT * FROM ${table} WHERE store_idx = ?`;
+            const storeIdxResult = await pool.queryParam_Arr(storeIdxQuery, [storeIdx]);
+
+            if (storeIdxResult[0] == undefined) {
+                resolve({
+                    code : statusCode.BAD_REQUEST,
+                    json : authUtil.successFalse(statusCode.BAD_REQUEST, `존재하지 않는 가게 펀드 정보(잘못된 인덱스)`)
+                });
+                return;
+            }
+
             const deleteStoreFundQuery = `DELETE FROM ${table} WHERE store_idx = ?`;
             const deleteStoreFundResult = await pool.queryParam_Arr(deleteStoreFundQuery, [storeIdx]);
